new build with shadowroot as default
