+++ src/ui/notifications/index.ts
@@ -1,3 +1,4 @@
+import './notification-container.component';
 import './notification.component';
 export { NotificationService } from './notification.service';
 export { NotificationType } from './notification.type';
