+++ src/ui/notifications/notification-container.component.scss
@@ -1,0 +1,22 @@
+@import '../../mixins';
+
+.notifications_wrapper {
+  position: fixed;
+  top: 3.5rem;
+  right: 0;
+  padding: 10px;
+}
+
+@include breakpoint(sm) {
+  .notifications_wrapper {
+    top: 0;
+    right: 0;
+    left: 0;
+    margin: 0 auto;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    z-index: 100;
+  }
+}
