+++ src/ui/notifications/notification.component.scss
@@ -1,83 +1,62 @@
 @import '../../mixins';
 
-$blue: #4A89DC;
+$blue: #4a89dc;
 $blue1: #3362cd;
 $blue-border: #005cbf;
-$red: #DA4453;
+$red: #da4453;
 $red1: #c82333;
 $red-border: #bd2130;
 $white: #fff;
 
-.notifications_wrapper {
-  position: fixed;
-  top: 3.5rem;
+.notification {
+  font-size: inherit;
+  font-family: inherit;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: relative;
+  margin-bottom: 10px;
   right: 0;
+  z-index: 1000;
   padding: 10px;
+  min-width: 15rem;
+  max-width: 18rem;
+  border-radius: 2px;
+  overflow: hidden;
+  box-shadow: 0 0.25rem 0.75rem rgba(0, 0, 0, 0.2);
 
-  .notification {
-    font-size: inherit;
-    font-family: inherit;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    position: relative;
-    margin-bottom: 10px;
-    right: 0;
-    z-index: 1000;
-    padding: 10px;
-    min-width: 15rem;
-    max-width: 18rem;
-    border-radius: 2px;
-    overflow: hidden;
-    box-shadow: 0 0.25rem 0.75rem rgba(0,0,0,.2);
-
-    &.is-info {
-      background-color: $blue;
-
-      .dismiss {
-        background-color: $blue1;
-        border-color: $blue-border;
-      }
-    }
-
-    &.is-danger {
-      background-color: $red;
-
-      .dismiss {
-        background-color: $red1;
-        border-color: $red-border;
-      }
-    }
-
-    &.is-info,
-    &.is-danger {
-      color: $white;
-    }
+  &.is-info {
+    background-color: $blue;
 
     .dismiss {
-      outline: 0;
-      border: none;
-      cursor: pointer;
-      border-radius: 2px;
-      color: $white;
-
-      &.hide-notify {
-        display: none;
-      }
+      background-color: $blue1;
+      border-color: $blue-border;
     }
   }
-}
 
-@include breakpoint(sm) {
-  .notifications_wrapper {
-    top: 0;
-    right: 0;
-    left: 0;
-    margin: 0 auto;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    z-index: 100;
+  &.is-danger {
+    background-color: $red;
+
+    .dismiss {
+      background-color: $red1;
+      border-color: $red-border;
+    }
+  }
+
+  &.is-info,
+  &.is-danger {
+    color: $white;
+  }
+
+  .dismiss {
+    outline: 0;
+    border: none;
+    cursor: pointer;
+    border-radius: 2px;
+    color: $white;
+
+    &.hide-notify {
+      display: none;
+    }
   }
 }
